Talents: Scroll, Potion and Treasure Bag Collector has been nerfed due to its nature being abused.

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
@@ -977,29 +977,29 @@ public abstract class Mob extends Char {
 				Dungeon.level.drop(Generator.random(Generator.Category.TREASUREBAG), pos).sprite.drop();
 			}
 
-			if (hero.pointsInTalent(Talent.SCROLL_COLLECTOR) >= 1) {
+			if (hero.pointsInTalent(Talent.SCROLL_COLLECTOR) >= 1 && Random.Float() < 0.2f) {
 				Dungeon.level.drop(Generator.random(Generator.Category.SCROLL), pos).sprite.drop();
-				if (hero.pointsInTalent(Talent.SCROLL_COLLECTOR) >= 2) {
+				if (hero.pointsInTalent(Talent.SCROLL_COLLECTOR) >= 2 && Random.Float() < 0.2f) {
 					Dungeon.level.drop(Generator.random(Generator.Category.SCROLL), pos).sprite.drop();
-					if (hero.pointsInTalent(Talent.SCROLL_COLLECTOR) >= 3) {
+					if (hero.pointsInTalent(Talent.SCROLL_COLLECTOR) >= 3 && Random.Float() < 0.2f) {
 						Dungeon.level.drop(Generator.random(Generator.Category.EXSCROLL), pos).sprite.drop();
 					}
 				}
 			}
 
-			if (hero.pointsInTalent(Talent.POTION_COLLECTOR) >= 1) {
+			if (hero.pointsInTalent(Talent.POTION_COLLECTOR) >= 1 && Random.Float() < 0.2f) {
 				Dungeon.level.drop(Generator.random(Generator.Category.POTION), pos).sprite.drop();
-				if (hero.pointsInTalent(Talent.POTION_COLLECTOR) >= 2) {
+				if (hero.pointsInTalent(Talent.POTION_COLLECTOR) >= 2 && Random.Float() < 0.2f) {
 					Dungeon.level.drop(Generator.random(Generator.Category.POTION), pos).sprite.drop();
-					if (hero.pointsInTalent(Talent.POTION_COLLECTOR) >= 3) {
+					if (hero.pointsInTalent(Talent.POTION_COLLECTOR) >= 3 && Random.Float() < 0.2f) {
 						Dungeon.level.drop(Generator.random(Generator.Category.EXPOTION), pos).sprite.drop();
 					}
 				}
 			}
 
-			if (hero.pointsInTalent(Talent.TREASUREBAG_COLLECTOR) >= 1) {
+			if (hero.pointsInTalent(Talent.TREASUREBAG_COLLECTOR) >= 1 && Random.Float() < 0.2f) {
 				Dungeon.level.drop(Generator.random(Generator.Category.TREASUREBAG), pos).sprite.drop();
-				if (hero.pointsInTalent(Talent.TREASUREBAG_COLLECTOR) >= 2) {
+				if (hero.pointsInTalent(Talent.TREASUREBAG_COLLECTOR) >= 2 && Random.Float() < 0.2f) {
 					Dungeon.level.drop(Generator.random(Generator.Category.TREASUREBAG), pos).sprite.drop();
 				}
 			}
